remove freezed & sanitize students/overview
